rename and package updates
