+++ CSharp/Util/Extensions/CollectionExtensions.cs
