+++ CSharp/Util/Internal/ResultCore.cs
@@ -1,0 +1,61 @@
+﻿//
+// Copyright © 2020 Terry Moreland
+// Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), 
+// to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, 
+// and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, 
+// WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+// 
+
+using System;
+using System.Diagnostics;
+
+namespace Moreland.CSharp.Util.Internal
+{
+    [DebuggerDisplay("{Success} {Message,nq}")]
+    internal struct ResultCore : IEquatable<ResultCore>
+    {
+        public ResultCore(bool success, string message, Exception? cause)
+        {
+            Success = success;
+            Message = message;
+            Cause = cause;
+        }
+
+        /// <summary>success or failure of which ever operation this represents the result for</summary>
+        public bool Success { get; }
+
+        /// <summary>Optional Message, should be non-empty on failure but may have a value on success as well</summary> 
+        public string Message { get; }
+
+        /// <summary>Optional additional exception detail</summary>
+        public Exception? Cause { get; }
+
+        public static bool operator ==(ResultCore leftHandSide, ResultCore rightHandSide) =>
+            leftHandSide.Equals(rightHandSide);
+        public static bool operator!=(ResultCore leftHandSide, ResultCore rightHandSide) =>
+            !(leftHandSide == rightHandSide);
+
+        #region IEquatable{ResultCore}
+        public bool Equals(ResultCore other) =>
+            Success == other.Success &&
+            Message == other.Message &&
+            (object.ReferenceEquals(Cause, other.Cause) || Cause?.Equals(other.Cause) == true);
+        #endregion
+        #region ValueType
+
+        public override bool Equals(object? obj) => obj is ResultCore result && Equals(result);
+        public override int GetHashCode() =>
+#           if NETSTANDARD2_0 || NET40 || NET45 || NET451 || NET452 || NET46 || NET461 || NET462 || NET47 || NET471 || NET472 || NET48
+            HashCodeBuilder.Create(Success, Message, Cause).ToHashCode();
+#           else
+            HashCode.Combine(Success, Message, Cause);
+#           endif
+
+        #endregion
+    }
+}
