--- DotNet/Util/Util.csproj
@@ -1,45 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netcoreapp3.1;netcoreapp2.1;netstandard2.0;netstandard2.1;net40;net45;net46;net472;net48</TargetFrameworks>
-    <RootNamespace>DotNet.Util</RootNamespace>
-    <AssemblyName>DotNet.Util</AssemblyName>
-    <LangVersion>8</LangVersion>
-    <Nullable>enable</Nullable>
-    <NeutralLanguage>en-US</NeutralLanguage>
-    <PackageLicenseFile></PackageLicenseFile>
-    <PackageId>DotNet.Util</PackageId>
-    <AssemblyVersion>1.0.0.0</AssemblyVersion>
-    <Version>1.0.0</Version>
-    <Authors>Terry Moreland</Authors>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <Company />
-    <Product>.NET Extension Library Utilities</Product>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <PackageDescription>Common Infrastructure classes</PackageDescription>
-    <RepositoryUrl>https://github.com/tsmoreland/Util</RepositoryUrl>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.CodeAnalysis.FxCopAnalyzers" Version="2.9.8">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Update="Properties\Resources.Designer.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>Resources.resx</DependentUpon>
-    </Compile>
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Update="Properties\Resources.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>Resources.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
-  </ItemGroup>
-
-</Project>
