+++ Util.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29728.190
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Util", "DotNet\Util\Util.csproj", "{1D6DC1E6-56D1-4A1D-8E91-227F1D85DA3B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Util", "CSharp\Util\Util.csproj", "{8A1707AE-7045-41BA-B5B8-1B94FA6E52F5}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Util.Test", "DotNet\Util.Test\Util.Test.csproj", "{CAB77004-D1D3-42FA-92C3-EFB93B13928F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Util.Test", "CSharp\Util.Test\Util.Test.csproj", "{41AF46E8-69A7-458A-9A1C-90ACFE4F6C5A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,14 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1D6DC1E6-56D1-4A1D-8E91-227F1D85DA3B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1D6DC1E6-56D1-4A1D-8E91-227F1D85DA3B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1D6DC1E6-56D1-4A1D-8E91-227F1D85DA3B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1D6DC1E6-56D1-4A1D-8E91-227F1D85DA3B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CAB77004-D1D3-42FA-92C3-EFB93B13928F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CAB77004-D1D3-42FA-92C3-EFB93B13928F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CAB77004-D1D3-42FA-92C3-EFB93B13928F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CAB77004-D1D3-42FA-92C3-EFB93B13928F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8A1707AE-7045-41BA-B5B8-1B94FA6E52F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8A1707AE-7045-41BA-B5B8-1B94FA6E52F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8A1707AE-7045-41BA-B5B8-1B94FA6E52F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8A1707AE-7045-41BA-B5B8-1B94FA6E52F5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{41AF46E8-69A7-458A-9A1C-90ACFE4F6C5A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{41AF46E8-69A7-458A-9A1C-90ACFE4F6C5A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{41AF46E8-69A7-458A-9A1C-90ACFE4F6C5A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{41AF46E8-69A7-458A-9A1C-90ACFE4F6C5A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
